+++ CHANGELOG.md
@@ -9,8 +9,10 @@ FlatLaf Change Log
   buttons.
 - TextComponent: Clip placeholder text if it does not fit into visible area. (PR
   #229)
-- Extras: Support embedding UI defaults inspector into any window (see
-  `FlatUIDefaultsInspector.createInspectorPanel()`.
+- Extras: UI defaults inspector:
+  - Support embedding UI defaults inspector panel into any window. See
+    `FlatUIDefaultsInspector.createInspectorPanel()`.
+  - Copy selected keys and values into clipboard via context menu.
 
 #### Fixed bugs